fix(Link): Use `currentTransitionStack` to enable auto-tracking (#394)
`currentURL` appears to update when the transition starts, but not anymore when it has ended. This leads to `isActive` not updating correctly if the `model()` hook is taking too long to load data. The `currentTransitionStack` however is changing in the `routeDidChange` event handler, which ensures that `isActive` (and friends) are updating at the right time.